remove noHeaders:true option, it makes form not read
I was trying to reverse this webapp, but due to
qr generation is using server sent event (SSE),
it does not work with http2. So i create custom path
in reverse proxy to another domain only for qr path.
It works but I need to define lots of hx-headers if
this option not enabled (CORS things)